+++ src/entry/common.ts
@@ -1,15 +1,8 @@
 export { FirmwareKind, Version } from '../headers';
 export { ParseError } from '../parse-error';
 export { Unit } from '../units';
+export { ParserEventKind } from '../data';
 
-export type {
-	ParserEvent,
-	ParserEventKind,
-	FrameFields,
-	MainFrame,
-	SlowFrame,
-	GpsFrame,
-	Stats,
-} from '../data';
+export type { ParserEvent, FrameFields, MainFrame, SlowFrame, GpsFrame, Stats } from '../data';
 export type { FrameDef } from '../headers';
 export type { WasmObject, WasmInit } from '../wasm';